MSVC8 Project Files + SDK Update

+++ dlls/csx/source/WinCSX/CRank.cpp
@@ -1,9 +1,7 @@
 #include "stdafx.h"
 
-//#include "amxxmodule.h"
 #include "CRank.h"
 #include <stdio.h>
-//#include "rank.h"
 
 // *****************************************************
 // class Stats
@@ -55,7 +53,7 @@ RankSystem::RankStats::~RankStats() {
 
 void RankSystem::RankStats::setName( const char* nn  )	{
 	delete[] name;
-	namelen = strlen(nn) + 1;
+	namelen = (short)strlen(nn) + 1;
 	name = new char[namelen];
 	if ( name )
 		strcpy( name , nn );
@@ -65,7 +63,7 @@ void RankSystem::RankStats::setName( const char* nn  )	{
 
 void RankSystem::RankStats::setUnique( const char* nn  )	{
 	delete[] unique;
-	uniquelen = strlen(nn) + 1;
+	uniquelen = (short)strlen(nn) + 1;
 	unique = new char[uniquelen];
 	if ( unique )
 		strcpy( unique , nn );
